refactor: move triggerDebugReport to debug.services to avoid circular dependencies.
